Move backend to separate directory
